+++ server-pg/init.sql
@@ -18,3 +18,8 @@ insert into configuration values ('device1', 100, 10, 54);
 insert into configuration values ('device2', 60, 0, 10);
 insert into configuration values ('device3', 60, 30, 40);
 insert into configuration values ('device4', 34, 10, 20);
+
+CREATE TABLE files (
+    name text,
+    file text
+);