+++ src/components/FormPanel/FormPanel.tsx
@@ -53,11 +53,13 @@ import {
 } from '../../types';
 import {
   convertToElementValue,
+  elementValueChangedCodeParameters,
   fileToBase64,
   getButtonVariant,
   getFieldValues,
   getInitialValuesMap,
   getPayloadForRequest,
+  requestCodeParameters,
   toFormData,
   toJson,
   ValueChangedEvent,
@@ -217,7 +219,7 @@ export const FormPanel: React.FC<Props> = ({
       code: string;
       initial: unknown;
       response?: FetchResponse | Response | null;
-      initialRequest?: () => void;
+      initialRequest?: () => Promise<void>;
       currentElements?: LocalFormElement[];
     }) => {
       if (!code) {
@@ -266,7 +268,7 @@ export const FormPanel: React.FC<Props> = ({
           notifySuccess,
           notifyWarning,
           toDataQueryResponse,
-          {
+          requestCodeParameters.create({
             grafana: {
               locationService,
               templateService: templateSrv,
@@ -295,7 +297,7 @@ export const FormPanel: React.FC<Props> = ({
               toDataQueryResponse,
               fileToBase64,
             },
-          }
+          })
         );
       } catch (error: unknown) {
         if (error instanceof Error) {
@@ -700,7 +702,13 @@ export const FormPanel: React.FC<Props> = ({
        * Set Content Type
        */
       const headers: HeadersInit = new Headers();
-      headers.set('Content-Type', options.update.contentType);
+
+      /**
+       * Browser should add header itself for form data content
+       */
+      if (options.update.contentType !== ContentType.FORMDATA) {
+        headers.set('Content-Type', options.update.contentType);
+      }
 
       /**
        * Set Header
@@ -823,43 +831,47 @@ export const FormPanel: React.FC<Props> = ({
     ({ elements, element }: { elements: LocalFormElement[]; element: LocalFormElement }) => {
       const fn = new Function('context', replaceVariables(options.elementValueChanged));
 
-      fn({
-        element,
-        grafana: {
-          locationService,
-          templateService: templateSrv,
-          notifyError,
-          notifySuccess,
-          notifyWarning,
-          eventBus,
-          appEvents,
-          refresh: () => appEvents.publish({ type: 'variables-changed', payload: { refreshAll: true } }),
-        },
-        panel: {
-          options,
-          data,
-          onOptionsChange,
-          elements,
-          onChangeElements,
-          initial: initialRef.current,
-          setError,
-          enableReset: () => setResetEnabled(true),
-          disableReset: () => setResetEnabled(false),
-          enableSubmit: () => setSubmitEnabled(true),
-          disableSubmit: () => setSubmitEnabled(false),
-          enableSaveDefault: () => setSaveDefaultEnabled(true),
-          disableSaveDefault: () => setSaveDefaultEnabled(false),
-        },
-        utils: {
-          toDataQueryResponse,
-        },
-      });
+      fn(
+        elementValueChangedCodeParameters.create({
+          element,
+          grafana: {
+            locationService,
+            templateService: templateSrv,
+            notifyError,
+            notifySuccess,
+            notifyWarning,
+            eventBus,
+            appEvents,
+            refresh: () => appEvents.publish({ type: 'variables-changed', payload: { refreshAll: true } }),
+          },
+          panel: {
+            options,
+            data,
+            onOptionsChange,
+            elements,
+            onChangeElements,
+            initial: initialRef.current,
+            setError,
+            enableReset: () => setResetEnabled(true),
+            disableReset: () => setResetEnabled(false),
+            enableSubmit: () => setSubmitEnabled(true),
+            disableSubmit: () => setSubmitEnabled(false),
+            enableSaveDefault: () => setSaveDefaultEnabled(true),
+            disableSaveDefault: () => setSaveDefaultEnabled(false),
+            initialRequest,
+          },
+          utils: {
+            toDataQueryResponse,
+          },
+        })
+      );
     },
     [
       appEvents,
       data,
       eventBus,
       initialRef,
+      initialRequest,
       notifyError,
       notifySuccess,
       notifyWarning,