+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 3.7.0 (IN PROGRESS)
+
+### Features / Enhancements
+
+- Add files upload examples and fix form data header (#357)
+- Add code parameters with builder and add initial request to element value changed code (#358)
+
 ## 3.6.0 (2023-01-10)
 
 ### Features / Enhancements